% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_volcano_plot.R
\name{create_all_volcano_plots}
\alias{create_all_volcano_plots}
\title{Create volcano plots for all comparisons in a DESeq2 object}
\usage{
create_all_volcano_plots(dds, ...)
}
\arguments{
\item{dds}{DESeqDataSet object}

\item{...}{Additional arguments passed to create_volcano_plot function}
}
\value{
A list of ggplot objects, one for each comparison
}
\description{
Create volcano plots for all comparisons in a DESeq2 object
}
\examples{
\dontrun{
all_plots <- create_all_volcano_plots(dds)
}
}
