% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gene_expr}
\alias{example_gene_expr}
\title{Example Gene Expression Data}
\format{
A data frame with 1000 rows and 6 variables:
\describe{
\item{gene_symbol}{Gene identifier}
\item{logFC}{Log2 fold change}
\item{P.Value}{Raw p-value}
\item{padj}{Adjusted p-value}
\item{log2FoldChange}{Log2 fold change (alternative format)}
\item{stat}{Test statistic}
}
}
\usage{
example_gene_expr
}
\description{
A dataset containing example differential expression results
}
\keyword{datasets}
