% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_gene_list}
\alias{example_gene_list}
\title{Example Gene List}
\format{
A character vector containing gene symbols
}
\usage{
example_gene_list
}
\description{
A vector of significant genes from the example gene expression data
}
\keyword{datasets}
