% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_qpcr_data}
\alias{example_qpcr_data}
\title{Example qPCR Data}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{Sample}{Sample name (Control or Treatment)}
\item{Target}{Target gene name (Target_Gene or Reference_Gene)}
\item{Cq}{Quantification cycle (Cq) values}
\item{Omit}{Logical indicating whether to omit the data point}
}
}
\usage{
example_qpcr_data
}
\description{
A dataset containing example qPCR data with control and treatment samples
}
\keyword{datasets}
