% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rfunk-package.R
\docType{package}
\name{Rfunk-package}
\alias{Rfunk}
\alias{Rfunk-package}
\title{Rfunk: Visualization Tools for Differential Expression Analysis}
\description{
A collection of useful R functions for bioinformatics analysis. Currently includes enhanced visualization tools for differential expression analysis, with features such as customizable volcano plots from DESeq2 results.
}
\author{
\strong{Maintainer}: SemiQuant \email{Jason.Limberis@ucsf.edu}

}
\keyword{internal}
