% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_signatures}
\alias{gene_signatures}
\title{Gene Signatures Dataset}
\format{
A data frame containing gene signatures
}
\source{
Internal gene signatures compilation
}
\usage{
gene_signatures
}
\description{
A dataset containing gene signatures information
}
\keyword{datasets}
